requirements.txt
robotframework-browser
robotframework==6.0
robotframework-pabot==2.13.0
browserstack-local
